Move out communication logic into a separate function from main
Now all of communication is in the peer-comm function outside of main.
To accomplish this, peer-comm uses the comm<message> effect handled
properly inside main (serializing/parsing in between), and a new effect
state which is just a getter/setter wrapper around a var of type
`peer-state` inside of main. This leads to a lot of ugly `get-state()`
instead of `state` code. This should probably be corrected at some
point.

Also inside this new function the logic for downloading consecutive
pieces is implemented, without saving any of it to a file, though.